+++ modules/tinymce/src/themes/silver/test/ts/browser/skin/OxideTablePickerMenuTest.ts
@@ -1,10 +1,11 @@
-import { ApproxStructure, Assertions, FocusTools, Keys, StructAssert } from '@ephox/agar';
+import { ApproxStructure, Assertions, FocusTools, Keys, Mouse, StructAssert, UiFinder, Waiter } from '@ephox/agar';
 import { describe, it } from '@ephox/bedrock-client';
 import { Fun } from '@ephox/katamari';
-import { SugarDocument } from '@ephox/sugar';
+import { SugarBody, SugarDocument } from '@ephox/sugar';
 import { TinyHooks, TinyUiActions } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
+import { Menu } from 'tinymce/core/api/ui/Ui';
 import Theme from 'tinymce/themes/silver/Theme';
 
 const tableCellsApprox = (s: ApproxStructure.StructApi, str: ApproxStructure.StringApi, arr: ApproxStructure.ArrayApi, selectedRows: number, selectedCols: number) => {
@@ -48,23 +49,30 @@ const insertTablePickerApprox = (s: ApproxStructure.StructApi, str: ApproxStruct
 
 describe('browser.tinymce.themes.silver.skin.OxideTablePickerMenuTest', () => {
   const hook = TinyHooks.bddSetup<Editor>({
+    menubar: 'table',
+    menu: {
+      table: { title: 'Table', items: 'table-menuitem' }
+    },
     toolbar: 'table-button',
     base_url: '/project/tinymce/js/tinymce',
     setup: (ed: Editor) => {
+      const tableMenuItem: Menu.FancyMenuItemSpec = {
+        type: 'fancymenuitem',
+        fancytype: 'inserttable',
+        onAction: Fun.noop
+      };
+
       ed.ui.registry.addMenuButton('table-button', {
-        type: 'menubutton',
-        fetch: (callback) => {
-          callback([
-            {
-              type: 'fancymenuitem',
-              fancytype: 'inserttable',
-              onAction: Fun.noop
-            }
-          ]);
-        }
+        icon: 'table',
+        fetch: (callback) => callback([ tableMenuItem ])
+      });
+
+      ed.ui.registry.addNestedMenuItem('table-menuitem', {
+        text: 'Insert table',
+        getSubmenuItems: () => [ tableMenuItem ]
       });
     }
-  }, [ Theme ]);
+  }, [ Theme ], true);
 
   it('TBA: Check structure of table picker', async () => {
     const editor = hook.editor();
@@ -85,5 +93,32 @@ describe('browser.tinymce.themes.silver.skin.OxideTablePickerMenuTest', () => {
       menu
     );
     await FocusTools.pTryOnSelector('Focus should be on 2 down, 2 across table cell', doc, '.tox-insert-table-picker__selected:last');
+    TinyUiActions.keydown(editor, Keys.escape());
+  });
+
+  it('TINY-6532: Re-opening the menu should reset the selected cells', async () => {
+    const editor = hook.editor();
+
+    const pOpenTablePicker = async () => {
+      const insertTableMenuItem = await TinyUiActions.pWaitForPopup(editor, '[role="menuitem"]:contains("Insert table")');
+      Mouse.mouseOver(insertTableMenuItem);
+      return await TinyUiActions.pWaitForPopup(editor, 'div.tox-fancymenuitem');
+    };
+
+    TinyUiActions.clickOnMenu(editor, 'button:contains("Table")');
+    const firstPicker = await pOpenTablePicker();
+    const item = UiFinder.findIn(firstPicker, 'div[role="button"]').getOrDie();
+    Mouse.mouseOver(item);
+    UiFinder.exists(firstPicker, 'div.tox-insert-table-picker__selected');
+    UiFinder.exists(firstPicker, 'span.tox-insert-table-picker__label:contains("1x1")');
+
+    TinyUiActions.keydown(editor, Keys.escape());
+    await Waiter.pTryUntil('Wait for menu to be hidden', () => UiFinder.notExists(SugarBody.body(), 'div.tox-fancymenuitem'));
+
+    const secondPicker = await pOpenTablePicker();
+    UiFinder.notExists(secondPicker, 'div.tox-insert-table-picker__selected');
+    UiFinder.exists(secondPicker, 'span.tox-insert-table-picker__label:contains("0x0")');
+    TinyUiActions.keydown(editor, Keys.escape());
+    TinyUiActions.keydown(editor, Keys.escape());
   });
 });